Fixed issue in ExampleShaderEditorApp when using RenderDoc.
Using RenderDoc with the shader editor example caused the 'cameraPos' uniform to be removed by shader optimization.
This triggered an exception in the rendering engine when trying to set the value of the uniform.
This commit fixes this by returning a boolean that is false when the uniform is not present in the shader.
The boolean is ignored when the uniform is not critical.