Added Ability to save nodes to file
Added Ability to save nodes to file, this can be adapted for a database as its a keyed json string that is generated.
I apologise that this is so many changes but I couldnt get it working without injecting some small changes into the code in relevant places, most are serialisation headers.
The key change is the addition of an Id and a ParentId, this allows the system to rebuild the data correctly.

Errors can occur if you try to deserialise and you have not added relevant serialisation headers or you have not set properties to get set properties.